requirements.txt
fastapi[standard]

huggingface_hub
onnxruntime_gpu==1.17.1
Shapely==2.0.5
numpy==1.26.4
Pillow==10.3.0
opencv_python==4.9.0.80
six==1.16.0
pyclipper==1.3.0.post5